refactor: TotalRecordsResponse 의 계산 로직을 AttendanceBook 으로 이동
