+++ src/main/java/controller/AttendanceController.java
@@ -1,5 +1,7 @@
 package controller;
 
+import static domain.AttendanceBook.fromAttendanceRecords;
+
 import domain.AttendanceBook;
 import domain.PenaltyStatus;
 import dto.AttendanceRecordResponse;
@@ -104,7 +106,7 @@ public class AttendanceController {
             String name = retryUntilValid(() -> askNameToCheckAttendance(attendanceBook));
 
             List<AttendanceRecordResponse> records = attendanceBook.getCrewByName(name).getAttendanceRecords();
-            TotalRecordsResponse totalRecord = TotalRecordsResponse.fromAttendanceRecords(records);
+            TotalRecordsResponse totalRecord = fromAttendanceRecords(records);
             PenaltyStatus penalty = PenaltyStatus.getByPenaltyCount(attendanceBook.getPenaltyCount(totalRecord));
             outputView.displayAttendanceRecordByName(name, records, totalRecord, penalty.getMessage());
         }
